requirements.txt
notebook
ipython
prefect
scikit-learn
openml
pyyaml
owlready2
rdflib
flask
pytest